moved stats saving to async thunk - preparing for introducing stat database layer
