+++ applications/launchpad/gui-react/src/store/containers/index.test.ts
@@ -1,7 +1,5 @@
-import { getEmptyStatsHistory } from '../../../__tests__/mocks/states/containersStateTemplates'
-
 import servicesReducer from './'
-import { SystemEventAction, ContainersState, Container } from './types'
+import { SystemEventAction, ContainersState } from './types'
 
 describe('store containers slice', () => {
   describe('updateStatus action', () => {
@@ -200,125 +198,6 @@ describe('store containers slice', () => {
           expect(nextState.stats.someContainerId.memory).toBe(0)
         })
       })
-    })
-  })
-
-  describe('stats', () => {
-    const standardTimestamp = '2022-05-23T12:13:14.000Z'
-    const standardStats = (read = standardTimestamp) => ({
-      read,
-      precpu_stats: {
-        cpu_usage: {
-          total_usage: 1,
-        },
-        system_cpu_usage: 1,
-      },
-      cpu_stats: {
-        cpu_usage: {
-          total_usage: 3,
-        },
-        system_cpu_usage: 5,
-        online_cpus: 1,
-      },
-      memory_stats: {
-        usage: 1024 * 1024,
-        stats: {},
-      },
-      networks: {},
-    })
-    const expectedFromStandardStats = (read = standardTimestamp) => ({
-      timestamp: read,
-      cpu: 50,
-      memory: 1,
-    })
-
-    it('should add stats under containerId', () => {
-      // given
-      const state = {
-        statsHistory: [],
-        pending: [],
-        containers: {
-          someContainerId: {
-            status: SystemEventAction.Start,
-          },
-        },
-        stats: {
-          someContainerId: {
-            cpu: 0,
-            memory: 0,
-          },
-        },
-      } as unknown as ContainersState
-      const expected = {
-        someContainerId: {
-          ...expectedFromStandardStats(),
-        },
-      }
-
-      // when
-      const nextState = servicesReducer(state, {
-        type: 'containers/stats',
-        payload: {
-          containerId: 'someContainerId',
-          service: Container.Tor,
-          stats: standardStats(),
-        },
-      })
-
-      // then
-      expect(nextState.stats).toStrictEqual(expected)
-    })
-
-    it('should retain stat history for service under current network, creating network state container if not yet available', () => {
-      // given
-      const state = {
-        statsHistory: [],
-        pending: [],
-        containers: {
-          someContainerId: {
-            status: SystemEventAction.Start,
-          },
-        },
-        stats: {
-          someContainerId: {
-            cpu: 0,
-            memory: 0,
-          },
-        },
-      } as unknown as ContainersState
-      const expected = {
-        pending: [],
-        containers: {
-          someContainerId: {
-            status: SystemEventAction.Start,
-          },
-        },
-        stats: {
-          someContainerId: {
-            ...expectedFromStandardStats(),
-          },
-        },
-        statsHistory: {
-          network: {
-            ...getEmptyStatsHistory(),
-            [Container.Tor]: [expectedFromStandardStats()],
-          },
-        },
-      }
-
-      // when
-      const nextState = servicesReducer(state, {
-        type: 'containers/stats',
-        payload: {
-          containerId: 'someContainerId',
-          service: Container.Tor,
-          network: 'network',
-          stats: standardStats(),
-        },
-      })
-
-      // then
-      expect(nextState).toStrictEqual(expected)
     })
   })
 })
