+++ applications/launchpad/gui-react/src/store/containers/index.ts
@@ -1,14 +1,12 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit'
 
 import {
-  StatsEventPayload,
   ContainerStatus,
   ContainersState,
-  ContainerId,
   Container,
   SystemEventAction,
 } from './types'
-import { start, stop, stopByType } from './thunks'
+import { addStats, start, stop, stopByType } from './thunks'
 
 const getInitialServiceStatus = (
   lastAction: SystemEventAction,
@@ -31,61 +29,12 @@ export const initialState: ContainersState = {
   pending: [],
   containers: {},
   stats: {},
-  statsHistory: {},
 }
 
 const servicesSlice = createSlice({
   name: 'containers',
   initialState,
   reducers: {
-    stats: (
-      state,
-      action: PayloadAction<{
-        containerId: ContainerId
-        stats: StatsEventPayload
-        service: Container
-        network: string
-      }>,
-    ) => {
-      const { stats, containerId, network, service } = action.payload
-
-      if (!state.stats || !state.stats[containerId]) {
-        return
-      }
-
-      const cs = stats.cpu_stats
-      const pcs = stats.precpu_stats
-      const cpu_delta = cs.cpu_usage.total_usage - pcs.cpu_usage.total_usage
-      const system_cpu_delta = cs.system_cpu_usage - pcs.system_cpu_usage
-      const numCpu = cs.online_cpus
-      const cpu = (cpu_delta / system_cpu_delta) * numCpu * 100.0
-
-      state.stats[containerId].timestamp = stats.read
-      state.stats[containerId].cpu = cpu
-
-      const ms = stats.memory_stats
-      const memory = (ms.usage - (ms.stats.cache || 0)) / (1024 * 1024)
-
-      state.stats[containerId].memory = memory
-
-      const { unsubscribe: _unsubscribe, ...statsToRemember } =
-        state.stats[containerId]
-      if (!state.statsHistory[network]) {
-        state.statsHistory = {
-          [network]: {
-            [Container.Tor]: [],
-            [Container.BaseNode]: [],
-            [Container.Wallet]: [],
-            [Container.SHA3Miner]: [],
-            [Container.MMProxy]: [],
-            [Container.XMrig]: [],
-            [Container.Monerod]: [],
-            [Container.Frontail]: [],
-          },
-        }
-      }
-      state.statsHistory[network][service].push(statsToRemember)
-    },
     updateStatus: (
       state,
       action: PayloadAction<{ containerId: string; action: SystemEventAction }>,
@@ -162,6 +111,18 @@ const servicesSlice = createSlice({
     builder.addCase(stop.rejected, (state, action) => {
       state.pending = state.pending.filter(p => p !== action.meta.arg)
       state.containers[action.meta.arg].error = action.payload
+    })
+
+    builder.addCase(addStats.fulfilled, (state, action) => {
+      const { containerId, stats } = action.payload
+
+      if (!state.stats || !state.stats[containerId]) {
+        return
+      }
+
+      state.stats[containerId].timestamp = stats.timestamp
+      state.stats[containerId].cpu = stats.cpu
+      state.stats[containerId].memory = stats.memory
     })
   },
 })
